Extract CursorPaginated concern from bot searches controller
